バッテリーセレクタのアイコン型を新規作成し、シンブレイバー攻撃アイコン生成関数を追加
