+++ src/js/game-object/battery-selector/view/battery-selector-icon/index.ts
@@ -1,4 +1,4 @@
-import { HorizontalAnimationMesh } from "../../../mesh/horizontal-animation";
+import { HorizontalAnimationMesh } from "../../../../mesh/horizontal-animation";
 
 /** バッテリーセレクタのアイコン */
 export type BatterySelectorIcon = {
